requirements.txt
# Core
browser-use>=0.1.0
playwright>=1.40.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# LLM Providers
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
